CHROMIUM: iwl7000: pcie: add pm_prepare and pm_complete ops
With these ops, we can know when we are about to enter system suspend.
This allows us to set a flag indicating that we are suspending and use
it to prevent the runtime resume operations (which is called as part
of the suspend flow) from grabbing the wakelock and aborting the
system suspend entry.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: a789886c8c15cc777ca6d98de4046be61788eca2