UPSTREAM: stmmac: fix resource management when resume
There is a memleak when suspend/resume this driver version.
Currently the stmmac, during resume step, reallocates all the resources
but they are not released when suspend.
The patch is not to release these resources but the logic has been changed.
In fact, it is not necessary to free and reallocate all from scratch
because the memory data will be always preserved.
As final solution, the patch just reinit the descriptors and the rx/tx
pointers only when resume. Tested done on STi boxes.

BUG=chrome-os-partner:56319
TEST=suspend_stress_test -c 5000

Reported-by: ZhengShunQian <zhengsq@rock-chips.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Cc: David S. Miller <davem@davemloft.net>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit ae79a639bb3dfd168dc8c1e5d6dfc471bdf6f284)
Signed-off-by: Shunqian Zheng <zhengsq@rock-chips.com>
Conflicts:
	drivers/net/ethernet/stmicro/stmmac/stmmac_main.c
Change-Id: I35e69d702c0e75ec7ad0aca89d04eee6f3a133a6
Reviewed-on: https://chromium-review.googlesource.com/389013
Tested-by: Ren Kuo <ren.kuo@quantatw.com>
Reviewed-by: Ren Kuo <ren.kuo@quantatw.com>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/422820