UPSTREAM: stmmac: create of compatible mdio bus for stmmac driver
The DSA driver needs to be passed a reference to an mdio bus. Typically
the mac is configured to use a fixed link but the mdio bus still needs
to be registered so that it con configure the switch.
This patch follows the same process as the altera tse ethernet driver for
creation of the mdio bus.

BUG=chrome-os-partner:57766
TEST=tested with upcoming patch

Acked-by: Rob Herring <robh@kernel.org>
Signed-off-by: Phil Reid <preid@electromag.com.au>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit e34d65696d2ef13dc32f2a162556c86c461ed763)
Change-Id: I68fe1498c047ddc47447a004016e2c1adac36cb9
Signed-off-by: Shunqian Zheng <zhengsq@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/412481
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>