CHROMIUM: iwl7000: xvt: add new tm API command
Add IWL_TM_USER_CMD_BUS_DATA_ACCESS command.
This command is already in use by windows driver and the USC.
Due to common API, it should be added also to the linux driver.
Though, in the linux driver it will not be supported, because
it's functionality will be implemented in the linux USC.

Signed-off-by: NitzanX Kugelman <nitzanx.kugelman@intel.com>
iwl7000-tree: 5d2b8dfc298ab20b135aaeb36df4f184e5ee2989