BACKPORT: clk: rockchip: handle mux dependency of fractional dividers
The fractional dividers of Rockchip SoCs contain an "auto-gating-feature"
that requires the downstream mux to actually point to the fractional
divider and the fractional divider gate to be enabled, for it to really
accept changes to the divider ratio.

The downstream muxes themselfs are not generic enough to include them
directly into the fractional divider, as they have varying sources of
parent clocks including not only clocks related to the fractional
dividers but other clocks as well.

To solve this, allow our clock branches to specify direct child clock-
branches in the new child property, let the fractional divider register
its downstream mux through this and add a clock notifier that temporarily
switches the mux setting when it notices rate changes to the fractional
divider.

NOTE:
  This is a significant backport.  The chromeos-3.14 kernel introduced
  commit 039c85487faec ("FROMLIST: clk: Add safe switch hook") which
  moved the clk post-rate-change notifications to a block after _all_
  clock rate changes were complete.

  This breaks the current special handling for fractional dividers, as
  the notifier that remuxes back to the original parent from the
  temporal one needs to run after the fractional rate change but before
  the mux rate gets handled. So to handle this without a notifier, we
  copy the fraction clk_ops, set a new set_rate callback that does this
  and calls the original set_rate in between.

  The clock-framework itself is supposed to get a completely new
  framework for handling all those clock-types depending on other
  clock-types (like our armclk and fractional dividers .. also Qualcomm
  cpuclk, etc) at some point in the future, so this is not meant for
  upstream, as hacking up the fractional clk_ops is pretty ugly and the
  original change that introduced the incompatibility also is one of
  those cases for the that new framework.

BUG=chrome-os-partner:48588, chrome-os-partner:34330
TEST=Minnie BT works w/ USB webcam plugged in.

Change-Id: Ia8265a5a41bfa8865f6acc8ebf4bb79b388bd81b
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Tested-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>
Reviewed-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/7907181/)
Reviewed-on: https://chromium-review.googlesource.com/320542
Tested-by: Lin Huang <hl@rock-chips.com>
Reviewed-by: Heiko Stübner <mmind00@googlemail.com>