CHROMIUM: clk: rockchip: never use USB PHY clock; SCLK_UART0 use fractional clock
Previously we were using the USB PHY clock as the source for UART0 to
get high speed UART.  However, it turns out that some usage of the USB
port (like hooking up to a webcam) causes that clock to be unsuitble to
use as the UART source clock.

This is probably the same thing that was described in
<http://crosreview.com/229554>, but at that time we believed the problem
to be a failing of the particular PHY associated with the EHCI port.
Now that we're seeing the same problem on the DWC2 ports it seems more
like a systemic problem.

Anyway, now that we've fixed the driver for the fractional clock we can
use it for the UART.  Hooray.

Upstream has done more with fractional clocks.  Since we're pretty
hopelessly different from upstream clock stuff, we'll just glom this
change in and change the one fractional clock that we need.  This also
pulls in a fix for a section mismatch that was fixed upstream in commit
5b73840375e3 ("clk: rockchip: fix section mismatches with new
child-clocks").

BUG=chrome-os-partner:48588, chrome-os-partner:34330
TEST=Minnie BT works w/ USB webcam plugged in.

Change-Id: I5971ff0f365c6695b2481e3d2d0f4929e4a2e93f
Signed-off-by: Lin Huang <hl@rock-chips.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/320543