UPSTREAM: Input: gpio-keys - fix check for disabling unsupported keys
Commit 4ea14a53d8f881034fa9e186653821c4e3d9a8fb ("Input: gpio-keys - report
error when disabling unsupported key") tried let user know that they
attempted to disable an unsupported key, unfortunately the check is wrong
as it believes that all codes are invalid. Fix it by ensuring that keys
that we try to disable are subset of keys (or switches) that device
reports.

Fixes: 4ea14a53d8f8 ("Input: gpio-keys - report error when disabling unsupported key")
Reported-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>
Tested-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>

BUG=b:71329519
TEST=Verify gpio-keys driver is loaded and bound to the eject gpio

(cherry picked from commit 8679ee4204cfd5cf78b996508ccadc1ec6130f1a)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: Ia51bb308faec7e05d742f8ea7f68fb710785e9e9
Reviewed-on: https://chromium-review.googlesource.com/890163
Reviewed-by: Rajat Jain <rajatja@chromium.org>