BACKPORT: PM / Domains: add setter for dev.pm_domain
Adds a function that sets the pointer to dev_pm_domain in struct device
and that warns if the device has already finished probing. The reason
why we want to enforce that is because in the general case that can
cause problems and also that we can simplify code quite a bit if we can
always assume that.

This patch also changes all current code that directly sets the
dev.pm_domain pointer.

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Reviewed-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 989561de9b5112999475b406557d9c7e9e59c041)

Conflicts:
	drivers/base/power/domain.c
	[due to earlier picked patch:
	  c7ca61734ba7 PM / Domains: defer dev_pm_domain_set() until genpd->attach_dev succeeds if present]

BUG=b:71779147
TEST=Test on scarlet, IOMMUs still attached

Change-Id: Ie0af8d90d1950d4a712f8311435fee5a6c98c84b
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/891041
Reviewed-by: Tomasz Figa <tfiga@chromium.org>