BACKPORT: PM / sleep: Go direct_complete if driver has no callbacks
If a suitable prepare callback cannot be found for a given device and
its driver has no PM callbacks at all, assume that it can go direct to
complete when the system goes to sleep.

The reason for this is that there's lots of devices in a system that do
no PM at all and there's no reason for them to prevent their ancestors
to do direct_complete if they can support it.

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Reviewed-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit aa8e54b559479d0cb7eb632ba443b8cacd20cd4b)

Conflicts:
	include/linux/pm.h
	[due to chromium dark resume patches]

BUG=b:71779147
TEST=Test on scarlet, IOMMUs still attached

Change-Id: I863963f5c0c48d3eaa690ce4ba55e4dbc9e513e9
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/870018
Reviewed-by: Tomasz Figa <tfiga@chromium.org>