FROMLIST: common: DMA-mapping: add DMA_ATTR_NOHUGEPAGE attribute
This patch adds the DMA_ATTR_NOHUGEPAGE attribute to the DMA-mapping
subsystem.

This attribute can be used as a hint to the DMA-mapping subsystem that
it's likely not worth it to try to allocation large pages behind the
scenes.  Large pages are likely to make an IOMMU TLB work more
efficiently but may not be worth it.  See the Documentation contained in
this patch for more details about this attribute and when to use it.

Note that the name of the hint (DMA_ATTR_NOHUGEPAGE) is based on the
name MADV_NOHUGEPAGE, which has the same meaning.  If we have expected
users, we could also add MADV_HUGEPAGE which has the opposite meaning of
this hint.

BUG=chromium:570532
TEST=Memory pressure + cat videos is even smoother!

Change-Id: I89968d51fbe2a98305bff4a01e0e0bc4ac9946d1
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Acked-by: Marek Szyprowski <m.szyprowski@samsung.com>
(am from https://patchwork.kernel.org/patch/7981221/)
Reviewed-on: https://chromium-review.googlesource.com/320496
Reviewed-by: Tomasz Figa <tfiga@chromium.org>