ANDROID: sched/tune: add support to compute normalized energy
The current EAS implementation considers only energy variations, while it
disregards completely the impact on performance for the selection of
a certain schedule candidate. Moreover, it also makes its decision based
on the "absolute" value of expected energy variations.

In order to properly define a trade-off strategy between increased energy
consumption and performances benefits it is required to compare energy
variations with performance variations.

Thus, both performance and energy metrics must be expressed in comparable
units. While the performance variations are expressed in terms of capacity
deltas, which are defined in the range [0..SCHED_LOAD_SCALE], the same
scale is not used for energy variations.

This patch introduces the function:
  schedtune_normalize_energy(energy_diff)
which returns a normalized value in the same range of capacity variations,
i.e. [0..SCHED_LOAD_SCALE].

A proper set of energy normalization constants are required to provide
a fast division by a constant during the normalziation of the energy_diff.
The value of these constants depends on the specific energy model and
topology of a target device.
Thus, this patch provides also the required support for the computation
at boot time of this set of variables.

BUG=chrome-os-partner:58667
TEST=build/boot on kevin

Change-Id: Icacc464dd32c84e2f7baca60bfcf79bcfc5d7ff4
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
(cherry picked from commit 637ee3715a1a22c07995be181c0e2fca10d2df2
 from android-msm/android-4.4.y)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/420036
(cherry picked from commit e5a1dba590fd0953a0accd07392424da3e68c874)
Reviewed-on: https://chromium-review.googlesource.com/424057