ANDROID: sched/fair: keep track of energy/capacity variations
The current EAS implementation does not allow "to boost" tasks
performances, for example by running them at an higher OPP (or a more
capable CPU), even if that could require a "reasonable" increase in
energy consumption.  To defined how much reasonable is an energy
increase with respect to a required boost value, it is required to
define and compute a trade-off between the expected energy and
performance variations.
However, the current EAS implementation considers only energy variations
while completely disregard the impact on performance for the selection
of a certain schedule candidate.

This patch extends the eenv energy environment to keep track of both
energy and performance deltas which are implied by the activation of a
schedule candidate.
The performance variation is estimated considering the different
capacities of the CPUs in which the task could be scheduled. The idea is
that while running on a CPU with higher capacity (e.g. higher operating
point) the task could (potentially) complete faster and thus get better
performance.

BUG=chrome-os-partner:58667
TEST=build/boot on kevin

Change-Id: I8cc079d24cc04d0dd36b6dc70cb24b257c620563
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
(cherry picked from commit 36967b2412fa7f356308c5104d95574e84ca03e
 from android-msm/android-4.4.y)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/420035
(cherry picked from commit 37f48b78ef344bf0b40e30bcf57b66583c9deb35)
Reviewed-on: https://chromium-review.googlesource.com/424133