CHROMIUM: drm: dw_hdmi: Add 68.25 and 75 to common_tmds_n_table
We missed these the first time for some reason.  Add them.  You can see
that the "n" in the table can work with an integer CTS to make the exact
frequency requested for 44.1kHz.

  >>> ((128 * 44100) * 65000) / 5376.
  68250000.0
  >>> ((128 * 44100) * 78125) / 5880.
  75000000.0

This change should be a no-op other than avoiding the slow calculation
in the case of the above two pixel clocks (and getting rid of the
warning from the logs).

BUG=chrome-os-partner:49514
TEST=Test audio at those clock rates

Change-Id: Ic391f3dad4b77b84e9f3197df0fadc8b6f731bc4
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/322436
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>