CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Decrease slop
We've now tightened up all of our sloppy HDMI rates by utilizing higher
Fvco source PLLs.  That means that we can undo the slop increase that we
did in http://crosreview.com/287296

It is possible that this will decrease the list of available resolutions
for some people, but the list presented should be much more solid.  This
seems like a worthwhile tradeoff, sepcially for mickey where a non-solid
HDMI mode can make the device appear dead.  If we find that there are
other rates that are needed then we should make exact versions of them
and not rely on slop.

This will "fix" problems seen with 1280x720 on one monitor.  On that
monitor we had the following mode:
  1280x720 120 1280 1376 1512 1744 720 721 724 772 flags: nhsync, pvsync

Specifically, 1744 * 772 * 120 = 161564160 is within the slop of
162000000 with the old slop but not the new:
  >>> 162000000 - 162000000/1000
  161838000
  >>> 162000000 - 162000000/333
  161513514

Losing "support" for the above mode line is probably a good thing since
apparently it didn't work reliably across all devices.

We don't add a replacement clock for the above mode since having 120 Hz
refresh rate will just eat up memory bandwidth for very little gain.  If
we found a non 120 Hz need for the above mode we can always add support
for it.

BUG=chrome-os-partner:49470
TEST=See that modes no longer work that required the old slop

Change-Id: I8de53e59355e131eb92c1238d8d1aa88c96d07b8
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/322846
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
(cherry picked from commit a202d0a6e4e615a702ae0a192306c482ab62f464)
Reviewed-on: https://chromium-review.googlesource.com/324115