UPSTREAM: crypto: hash - Add helpers to zero stack request/descriptor
As the size of an ahash_request or shash_desc is variable, it's
awkward to zero them explicitly.  This patch adds helpers to do
that which should be used when they are created on the stack.

BUG=chromium:704196
TEST=Deploy v5.44 BlueZ system daemon with v4.9 BlueZ kernel on Oak and
perform automated tests (suite:bluetooth) and manual tests against mouses,
keyboards and headphones.

Change-Id: Id32411757f8172706d7d97b98c7b5ac1ae406910
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Miao-chen Chou <mcchou@chromium.org>
(cherry picked from commit e67ffe0af4d450e95e5f9bb7d081500e30777857)
Reviewed-on: https://chromium-review.googlesource.com/474083
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>