UPSTREAM: crypto: skcipher - Add helper to zero stack request
As the size of an skcipher_request is variable, it's awkward to
zero it explicitly.  This patch adds a helper to do that which
should be used when it is created on the stack.

BUG=chromium:704196
TEST=Deploy v5.44 BlueZ system daemon with v4.9 BlueZ kernel on Minnie and
perform automated tests (suite:bluetooth) and manual tests against mouses,
keyboards and headphones.

Change-Id: I8937e43c39440bb5fee3927bf7e49abf9bf5fdd3
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Miao-chen Chou <mcchou@chromium.org>
(cherry picked from commit 1aaa753d918c48c603195a468766e6a2b32b87f9)
Reviewed-on: https://chromium-review.googlesource.com/475324
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>