CHROMIUM: Still set backlight level if pipe is invalid
This moves the invalid check for the connector pipe down to when the
backlight is actually set. This fixes the case where backlight changes are
ignored if the modeset for the panel is in progress. The backlight device
would think the backlight level was one value while another value was
actually used for the panel.

BUG=chromium:535021
TEST=suspend/resume on link
check the backlight level isn't changed after resume

Change-Id: I4d935fcb60aa824da15eb79d7f8567ed34c23877
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/323407
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>