UPSTREAM: ARM: 8503/1: clk_register_clkdev: remove format string interface
Many callers either use NULL or const strings for the third argument of
clk_register_clkdev. For those that do not and use a non-const string,
this is a risk for format strings being accidentally processed (for
example in device names). As this interface is already used as if it
weren't a format string (prints nothing when NULL), and there are zero
users of the format strings, remove the format string interface to make
sure format strings will not leak into the clkdev.

$ git grep '\bclk_register_clkdev\b' | grep % | wc -l
0

Unfortunately, all the internals expect a va_list even though they treat
a NULL format string as special. To deal with this, we must pass either
(..., "%s", string) or (..., NULL) so that a the va_list will be created
correctly (passing the name as an argument, not as a format string).

Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
(cherry picked from commit 416dd13ad620a14fbabe5d73584b12e07ce8d02e)
Signed-off-by: Hsin-Yu Chao <hychao@chromium.org>

BUG=b:62106805
TEST=Apply full patch set, verify capture alone can get valid data
and playback doesn't have pop noise.

Change-Id: I20ca6885433b5c992a6118f42790464ac910d398
Reviewed-on: https://chromium-review.googlesource.com/676508
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>