UPSTREAM: mmc: sdhci: avoid unnecessary mapping/unmapping of align buffer
Unnecessarily mapping and unmapping the align buffer for SD cards is
expensive: performance measurements on iMX6 show that this gives a hit
of 10% on hdparm buffered disk reads.

MMC/SD card IO comes from the mm/vfs which gives us page based IO, so
for this case, the align buffer is not going to be used.  However, we
still map and unmap this buffer.

Eliminate this by switching the align buffer to be a DMA coherent
buffer, which needs no DMA maintenance to access the buffer.

BUG=chromium:721944
TEST=hardware_StorageFio

Change-Id: I25bb969044cf5ef4ced7f0f8b57b9bc3eccbb4b4
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Cc: stable@vger.kernel.org # v4.5+
Tested-by: Gregory CLEMENT <gregory.clement@free-electrons.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit edd63fcc97cdb53279a7c43fa1691f5913d92793)
Reviewed-on: https://chromium-review.googlesource.com/508008
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>